:seedling: Initialize ClimSim Burn model
Initial implementation of the ClimSim neural network model implemented using Burn! The dataloader only loads the state_u_0 and ptend_u_0 columns as input and target for now, and the model itself is simply two linear layers with a ReLU activation. Also updated the main README.md with instructions to run the model, including a GIF animation of the TUI training dashboard.